Update 简体中文
I retranslated the latest version of the text. I really hope to see the language of my country soon. Can I advertise this modifier on the video platform after the update? I will indicate the author, the ownership of the author and the original address of this modifier. I hope you can agree. My English is not very good, if there are some grammatical errors, please forgive me! I really mean no harm.