Add files via upload
I'm sorry, it's my first time to participate in translation, which has caused you a little trouble. I have corrected it again. Because my English is not very good, some texts have not been translated. I don't know if I can upload some of them in the way of translation, and keep the rest in the original.