Add support for Chinese
* Create 简体中文

* Update 简体中文

I retranslated the latest version of the text. I really hope to see the language of my country soon. Can I advertise this modifier on the video platform after the update? I will indicate the author, the ownership of the author and the original address of this modifier. I hope you can agree. My English is not very good, if there are some grammatical errors, please forgive me! I really mean no harm.

* Add files via upload

I'm sorry, it's my first time to participate in translation, which has caused you a little trouble. I have corrected it again. Because my English is not very good, some texts have not been translated. I don't know if I can upload some of them in the way of translation, and keep the rest in the original.